Added Batch Number, Fill Weights and Auto Email on Order Creation
Added batch number to fillings and form remembers previous batch number for ease of use. Added fill weight to filling which is end weight - tare weight, and total fill weight to order which is the sum of all of its fillings' fill weights. Now an email sends to the email address in secret.py when an order is created, with a generic message. Fixed some navigation issues.