Improved Order Form and Cylinder Expiry
Create order page is aligned as a table, while fill type and stillage are removed, field instruction is changed to QC instruction, products are now perfluorobutane and octafluoropropane, and cylinder size is now a number input. Cylinder list shows when a cylinder is close to expiry and expired including highlighting as yellow and red respectively. Fixed add new order line, entering a cylinderid with letters, and improved email formatting. Also better navigation including getting rid of the menu page and making order list the main page.